Getting ready to hide objects
Now, since our camera is in one locked position, there is only one wall that blocks our vision. The easy solution would be just remove that wall but it would result in the level looking weird.

This solution should use raycasts and make the wall transparent when the camera is behind it.